System crashes if there is process migration during vfp_init() call.
During vfp_init(), if a process which called vfp_enable() is migrated
just after the call, then the process executing the rest of code will
access a VFP unit which is not ENABLED and also smp_call_function() will
not work as it is expected.

This patch prevents accessing VFP unit disabled by preventing migration
and also replaces smp_call_function() with on_each_cpu() to make sure
that no VFP remains disabled.

Signed-off-by: Hyungwoo Yang <hyungwooy@nvidia.com>