MMC tweaks
1) mmc: card: increase command retries to 10 from 3
When commands timeout, previously we had code to
retry the same command 3 times. But under some
situations 3 retries do not suffice. Increasing
the retries to 10 does the trick. Also if the card
does not respond after 10 retries then the card is
dead for sure. But if the same card responds in
between 3 to 10 retries then it is always beneficial
to have retries as 10.
Bug 914934
Varun Wadekar <vwadekar@nvidia.com>

2) Remove Asus debug spamming during normal operations