Implement CONFIG_ZIMMER by HTC Studio Engineering (reference OneX kernel)
Experimental - can't find the original source of this hack, so it may
be an HTC original.

-Increases vmscan swappiness to 100 by default to force swapping of
bloated/dead RAM to open up real RAM for something more useful. This
will be interesting to test with ZRAM to see if it is complimentary. 
swappiness can still be changed via sysfs if desired.

-Implements tweaks to block, fs, and mm to flag and handle i/o requests
differently that are involved in swap or are ready to be moved to
swap.